APIGOV-24312 - Fix tests
