+++ pkg/apic/auth/apicauth_test.go
@@ -45,74 +45,6 @@ func TestChannelTokenGetterCloses(t *testing.T) {
 	}
 }
 
-func TestGetKey(t *testing.T) {
-	cases := []struct {
-		description string
-		kr          *keyReader
-	}{
-		{
-			"no password",
-			&keyReader{
-				privKey: "testdata/private_key.pem",
-			},
-		},
-		{
-			"with empty password file",
-			&keyReader{
-				privKey:  "testdata/private_key.pem",
-				password: "testdata/password_empty",
-			},
-		},
-		{
-			"with password",
-			&keyReader{
-				privKey:  "testdata/private_key_with_pwd.pem",
-				password: "testdata/password",
-			},
-		},
-	}
-
-	for _, testCase := range cases {
-		if _, err := testCase.kr.getPrivateKey(); err != nil {
-			t.Errorf("testcase: %s: failed to read rsa key %s", testCase.description, err)
-		}
-	}
-}
-
-func TestGetPublicKey(t *testing.T) {
-	cases := []struct {
-		description string
-		kr          *keyReader
-	}{
-		{
-			"with public key",
-			&keyReader{
-				publicKey: "testdata/public_key",
-			},
-		},
-		{
-			"with private and public key",
-			&keyReader{
-				privKey:   "testdata/private_key.pem",
-				publicKey: "testdata/public_key",
-			},
-		},
-		{
-			"with private, public key, and password",
-			&keyReader{
-				privKey:   "testdata/private_key_with_pwd.pem",
-				password:  "testdata/password",
-				publicKey: "testdata/public_key",
-			},
-		},
-	}
-	for _, testCase := range cases {
-		if _, err := testCase.kr.getPublicKey(); err != nil {
-			t.Errorf("testcase: %s: failed to read public key %s", testCase.description, err)
-		}
-	}
-}
-
 func TestNetAuthenticate(t *testing.T) {
 	aa := NewWithStatic("12345", "abcde")
 	if aa.tenantID != "12345" {
